refactor inner reducers to their own file
