+++ core/src/dht/dht_reducers.rs
@@ -6,22 +6,15 @@ use crate::{
     dht::dht_store::DhtStore,
     network::entry_with_header::EntryWithHeader,
 };
-use holochain_core_types::{
-    cas::content::{Address, AddressableContent},
-    crud_status::{create_crud_link_eav, create_crud_status_eav, CrudStatus},
-    eav::{Attribute, EaviQuery, EntityAttributeValueIndex, IndexFilter},
-    entry::Entry,
-    error::{HcResult, HolochainError},
-    link::Link,
+use std::sync::Arc;
+
+use super::dht_inner_reducers::{
+    reduce_add_remove_link_inner, reduce_remove_entry_inner, reduce_store_entry_inner,
+    reduce_update_entry_inner, LinkModification,
 };
-use std::{collections::BTreeSet, convert::TryFrom, str::FromStr, sync::Arc};
 
 // A function that might return a mutated DhtStore
 type DhtReducer = fn(Arc<Context>, &DhtStore, &ActionWrapper) -> Option<DhtStore>;
-enum LinkModification {
-    Add,
-    Remove,
-}
 
 /// DHT state-slice Reduce entry point.
 /// Note: Can't block when dispatching action here because we are inside the reduce's mutex
@@ -31,15 +24,14 @@ pub fn reduce(
     action_wrapper: &ActionWrapper,
 ) -> Arc<DhtStore> {
     // Get reducer
-    let maybe_reducer = resolve_reducer(action_wrapper);
-    if maybe_reducer.is_none() {
-        return old_store;
-    }
-    let reducer = maybe_reducer.unwrap();
+    let reducer = match resolve_reducer(action_wrapper) {
+        Some(reducer) => reducer,
+        None => {
+            return old_store;
+        }
+    };
     // Reduce
-    let store = old_store.clone();
-    let maybe_new_store = reducer(context, &store, &action_wrapper);
-    match maybe_new_store {
+    match reducer(context, &old_store.clone(), &action_wrapper) {
         None => old_store,
         Some(new_store) => Arc::new(new_store),
     }
@@ -93,32 +85,15 @@ pub(crate) fn reduce_hold_entry(
     }
 }
 
-/// This is used as the inner function for both commit and hold reducers
-fn reduce_store_entry_inner(store: &mut DhtStore, entry: &Entry) -> HcResult<()> {
-    match (*store.content_storage().write()?).add(entry) {
-        Ok(()) => create_crud_status_eav(&entry.address(), CrudStatus::Live).map(|status_eav| {
-            (*store.meta_storage().write().unwrap())
-                .add_eavi(&status_eav)
-                .map(|_| ())
-                .map_err(|e| format!("err/dht: dht::reduce_hold_entry() FAILED {:?}", e).into())
-        })?,
-        Err(e) => Err(format!("err/dht: dht::reduce_hold_entry() FAILED {:?}", e).into()),
-    }
-}
-
 pub(crate) fn reduce_add_link(
     _context: Arc<Context>,
     old_store: &DhtStore,
     action_wrapper: &ActionWrapper,
 ) -> Option<DhtStore> {
-    // Get Action's input data
     let link = unwrap_to!(action_wrapper.action() => Action::AddLink);
     let mut new_store = (*old_store).clone();
-
     let res = reduce_add_remove_link_inner(&mut new_store, link, LinkModification::Add);
-
     new_store.actions_mut().insert(action_wrapper.clone(), res);
-
     Some(new_store)
 }
 
@@ -127,62 +102,11 @@ pub(crate) fn reduce_remove_link(
     old_store: &DhtStore,
     action_wrapper: &ActionWrapper,
 ) -> Option<DhtStore> {
-    // Get Action's input data
     let link = unwrap_to!(action_wrapper.action() => Action::RemoveLink);
     let mut new_store = (*old_store).clone();
-
     let res = reduce_add_remove_link_inner(&mut new_store, link, LinkModification::Remove);
-
     new_store.actions_mut().insert(action_wrapper.clone(), res);
-
     Some(new_store)
-}
-
-fn reduce_add_remove_link_inner(
-    store: &mut DhtStore,
-    link: &Link,
-    link_modification: LinkModification,
-) -> HcResult<Address> {
-    match (*store.content_storage().read()?).contains(link.base())? {
-        true => {
-            let attr = match link_modification {
-                LinkModification::Add => Attribute::LinkTag(link.tag().to_string()),
-                LinkModification::Remove => Attribute::RemovedLink(link.tag().to_string()),
-            };
-            let eav = EntityAttributeValueIndex::new(link.base(), &attr, link.target())?;
-            store.meta_storage().write()?.add_eavi(&eav)?;
-            Ok(link.base().clone())
-        }
-        false => Err(HolochainError::ErrorGeneric(String::from(
-            "Base for link not found",
-        ))),
-    }
-}
-
-fn reduce_update_entry_inner(
-    store: &DhtStore,
-    old_address: &Address,
-    new_address: &Address,
-) -> HcResult<Address> {
-    // Update crud-status
-
-    let new_status_eav = create_crud_status_eav(old_address, CrudStatus::Modified)?;
-    (*store.meta_storage().write()?).add_eavi(&new_status_eav)??;
-
-    let crud_link_eav = create_crud_link_eav(old_address, new_address)?;
-    (*store.meta_storage().write()?).add_eavi(&crud_link_eav)??;
-
-    Ok(new_address.clone())
-
-    // match (*store.meta_storage().write()?).add_eavi(&new_status_eav)? {
-    //     Some(_) => Ok(new_address.clone()),
-    //     None => {
-    //         match (*store.meta_storage().write()?).add_eavi(&crud_link_eav)? {
-    //             Some(_) => Ok(new_address.clone()),
-    //             None => Err("wtf".into())
-    //         }
-    //     }
-    // }
 }
 
 pub(crate) fn reduce_update_entry(
@@ -190,13 +114,9 @@ pub(crate) fn reduce_update_entry(
     old_store: &DhtStore,
     action_wrapper: &ActionWrapper,
 ) -> Option<DhtStore> {
-    // Setup
-    let action = action_wrapper.action();
-    let (old_address, new_address) = unwrap_to!(action => Action::UpdateEntry);
+    let (old_address, new_address) = unwrap_to!(action_wrapper.action() => Action::UpdateEntry);
     let mut new_store = (*old_store).clone();
-    // Act
     let res = reduce_update_entry_inner(&mut new_store, old_address, new_address);
-    // Done
     new_store.actions_mut().insert(action_wrapper.clone(), res);
     Some(new_store)
 }
@@ -206,85 +126,14 @@ pub(crate) fn reduce_remove_entry(
     old_store: &DhtStore,
     action_wrapper: &ActionWrapper,
 ) -> Option<DhtStore> {
-    // Setup
-    let action = action_wrapper.action();
-    let (deleted_address, deletion_address) = unwrap_to!(action => Action::RemoveEntry);
+    let (deleted_address, deletion_address) =
+        unwrap_to!(action_wrapper.action() => Action::RemoveEntry);
     let mut new_store = (*old_store).clone();
-    // Act
     let res = reduce_remove_entry_inner(&mut new_store, deleted_address, deletion_address);
-    // Done
     new_store.actions_mut().insert(action_wrapper.clone(), res);
     Some(new_store)
 }
 
-//
-fn reduce_remove_entry_inner(
-    new_store: &mut DhtStore,
-    latest_deleted_address: &Address,
-    deletion_address: &Address,
-) -> HcResult<Address> {
-    // pre-condition: Must already have entry in local content_storage
-    let content_storage = &new_store.content_storage().clone();
-
-    let maybe_json_entry = content_storage
-        .read()
-        .unwrap()
-        .fetch(latest_deleted_address)
-        .unwrap();
-    let json_entry = maybe_json_entry.ok_or_else(|| {
-        HolochainError::ErrorGeneric(String::from("trying to remove a missing entry"))
-    })?;
-
-    let entry = Entry::try_from(json_entry).expect("Stored content should be a valid entry.");
-    // pre-condition: entry_type must not by sys type, since they cannot be deleted
-    if entry.entry_type().to_owned().is_sys() {
-        return Err(HolochainError::ErrorGeneric(String::from(
-            "trying to remove a system entry type",
-        )));
-    }
-    // pre-condition: Current status must be Live
-    // get current status
-    let meta_storage = &new_store.meta_storage().clone();
-    let status_eavs = meta_storage.read().unwrap().fetch_eavi(&EaviQuery::new(
-        Some(latest_deleted_address.clone()).into(),
-        Some(Attribute::CrudStatus).into(),
-        None.into(),
-        IndexFilter::LatestByAttribute,
-    ))?;
-
-    //TODO clean up some of the early returns in this
-    // TODO waiting for update/remove_eav() assert!(status_eavs.len() <= 1);
-    // For now checks if crud-status other than Live are present
-    let status_eavs = status_eavs
-        .into_iter()
-        .filter(|e| CrudStatus::from_str(String::from(e.value()).as_ref()) != Ok(CrudStatus::Live))
-        .collect::<BTreeSet<EntityAttributeValueIndex>>();
-    if !status_eavs.is_empty() {
-        return Err(HolochainError::ErrorGeneric(String::from(
-            "entry_status != CrudStatus::Live",
-        )));
-    }
-    // Update crud-status
-    let result = create_crud_status_eav(latest_deleted_address, CrudStatus::Deleted);
-    if result.is_err() {
-        return Err(HolochainError::ErrorGeneric(String::from(
-            "Could not create eav",
-        )));
-    }
-    let new_status_eav = result.expect("should unwrap eav");
-    let meta_storage = &new_store.meta_storage().clone();
-
-    (*meta_storage.write().unwrap()).add_eavi(&new_status_eav)?;
-
-    // Update crud-link
-    let crud_link_eav = create_crud_link_eav(latest_deleted_address, deletion_address)
-        .map_err(|_| HolochainError::ErrorGeneric(String::from("Could not create eav")))?;
-    let res = (*meta_storage.write().unwrap()).add_eavi(&crud_link_eav);
-
-    res.map(|_| latest_deleted_address.clone())
-}
-
-//
 #[allow(dead_code)]
 pub(crate) fn reduce_get_links(
     _context: Arc<Context>,
