+++ core/src/dht/mod.rs
@@ -3,3 +3,5 @@
 pub mod actions;
 pub mod dht_reducers;
 pub mod dht_store;
+
+mod dht_inner_reducers;
